# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MCP Gateway Registry",
    "description": "A registry and management system for Model Context Protocol (MCP) servers",
    "version": "1.0.11"
  },
  "paths": {
...
    "/api/federation/config": {
      "get": {
        "tags": ["federation", "federation"],
        "summary": "Get federation configuration",
        "description": "Get federation configuration by ID.\n\nArgs:\n    config_id: Configuration ID (default: \"default\")\n    user_context: Authenticated user context\n    repo: Federation config repository\n\nReturns:\n    Federation configuration\n\nRaises:\n    404: Configuration not found",
        "operationId": "get_federation_config_api_federation_config_get",
        "parameters": [
          {
            "name": "config_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "default",
              "title": "Config Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Federation Config Api Federation Config Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "post": {
        "tags": ["federation", "federation"],
        "summary": "Create or update federation configuration",
        "description": "Create or update federation configuration.\n\nArgs:\n    config: Federation configuration to save\n    config_id: Configuration ID (default: \"default\")\n    user_context: Authenticated user context\n    repo: Federation config repository\n\nReturns:\n    Saved configuration\n\nExample:\n    ```json\n    {\n      \"anthropic\": {\n        \"enabled\": true,\n        \"endpoint\": \"https://registry.modelcontextprotocol.io\",\n        \"sync_on_startup\": false,\n        \"servers\": [\n          {\"name\": \"io.github.jgador/websharp\"},\n          {\"name\": \"modelcontextprotocol/filesystem\"}\n        ]\n      },\n      \"asor\": {\n        \"enabled\": false,\n        \"endpoint\": \"\",\n        \"auth_env_var\": \"ASOR_ACCESS_TOKEN\",\n        \"sync_on_startup\": false,\n        \"agents\": []\n      }\n    }\n    ```",
        "operationId": "save_federation_config_api_federation_config_post",
        "parameters": [
          {
            "name": "config_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "default",
              "title": "Config Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FederationConfig"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Save Federation Config Api Federation Config Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/federation/config/{config_id}": {
      "put": {
        "tags": ["federation", "federation"],
        "summary": "Update specific federation configuration",
        "description": "Update a specific federation configuration.\n\nArgs:\n    config_id: Configuration ID to update\n    config: Updated federation configuration\n    user_context: Authenticated user context\n    repo: Federation config repository\n\nReturns:\n    Updated configuration",
        "operationId": "update_federation_config_api_federation_config__config_id__put",
        "parameters": [
          {
            "name": "config_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Config Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FederationConfig"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Federation Config Api Federation Config  Config Id  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "tags": ["federation", "federation"],
        "summary": "Delete federation configuration",
        "description": "Delete a federation configuration.\n\nArgs:\n    config_id: Configuration ID to delete\n    user_context: Authenticated user context\n    repo: Federation config repository\n\nReturns:\n    Deletion confirmation\n\nRaises:\n    404: Configuration not found",
        "operationId": "delete_federation_config_api_federation_config__config_id__delete",
        "parameters": [
          {
            "name": "config_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Config Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Delete Federation Config Api Federation Config  Config Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/federation/configs": {
      "get": {
        "tags": ["federation", "federation"],
        "summary": "List all federation configurations",
        "description": "List all federation configurations.\n\nArgs:\n    user_context: Authenticated user context\n    repo: Federation config repository\n\nReturns:\n    List of configuration summaries with id, created_at, updated_at",
        "operationId": "list_federation_configs_api_federation_configs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response List Federation Configs Api Federation Configs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/federation/config/{config_id}/anthropic/servers": {
      "post": {
        "tags": ["federation", "federation"],
        "summary": "Add Anthropic server to config",
        "description": "Add a server to Anthropic federation configuration.\n\nArgs:\n    config_id: Configuration ID\n    server_name: Server name to add (e.g., \"io.github.jgador/websharp\")\n    user_context: Authenticated user context\n    repo: Federation config repository\n\nReturns:\n    Updated configuration",
        "operationId": "add_anthropic_server_api_federation_config__config_id__anthropic_servers_post",
        "parameters": [
          {
            "name": "config_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Config Id"
            }
          },
          {
            "name": "server_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Server Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Add Anthropic Server Api Federation Config  Config Id  Anthropic Servers Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/api/federation/config/{config_id}/asor/agents": {
      "post": {
        "tags": ["federation", "federation"],
        "summary": "Add ASOR agent to config",
        "description": "Add an agent to ASOR federation configuration.\n\nArgs:\n    config_id: Configuration ID\n    agent_id: Agent ID to add (e.g., \"aws_assistant\")\n    user_context: Authenticated user context\n    repo: Federation config repository\n\nReturns:\n    Updated configuration",
        "operationId": "add_asor_agent_api_federation_config__config_id__asor_agents_post",
        "parameters": [
          {
            "name": "config_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Config Id"
            }
          },
          {
            "name": "agent_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Add Asor Agent Api Federation Config  Config Id  Asor Agents Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FederationConfig": {
        "properties": {
          "anthropic": {
            "$ref": "#/components/schemas/AnthropicFederationConfig"
          },
          "asor": {
            "$ref": "#/components/schemas/AsorFederationConfig"
          }
        },
        "type": "object",
        "title": "FederationConfig",
        "description": "Root federation configuration."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer token obtained from Keycloak OAuth2 authentication. Include in Authorization header as: `Authorization: Bearer <token>`"
      }
    }
  }
}